Linux compile fixes

+++ ApplicationCode/ProjectDataModel/ProjectDataModel_UnitTests/RimWellLogExtractionCurveImpl-Test.cpp
@@ -45,9 +45,9 @@ TEST(RimWellLogExtractionCurveImplTest, StripOffHugeValAtEndsAndInteriorOfVector
     std::vector< std::pair<size_t, size_t> > valuesIntervals;
     RigWellLogCurveDataTestInterface::calculateIntervalsOfValidValues(values, &valuesIntervals);
 
-    EXPECT_EQ(2, valuesIntervals.size());
-    EXPECT_EQ(2, valuesIntervals[0].first);
-    EXPECT_EQ(2, valuesIntervals[0].second);
-    EXPECT_EQ(5, valuesIntervals[1].first);
-    EXPECT_EQ(6, valuesIntervals[1].second);
+    EXPECT_EQ(2, static_cast<int>(valuesIntervals.size()));
+    EXPECT_EQ(2, static_cast<int>(valuesIntervals[0].first));
+    EXPECT_EQ(2, static_cast<int>(valuesIntervals[0].second));
+    EXPECT_EQ(5, static_cast<int>(valuesIntervals[1].first));
+    EXPECT_EQ(6, static_cast<int>(valuesIntervals[1].second));
 }
